driver ping pong reorganized
