--- linux_drivers/MIG_2_DDR3_DMA/defines_dma.h
@@ -1,20 +1,0 @@
-#define MAP_SIZE 0x40000
-
-#define BASE_ADDRES 0
-
-#define XPREDICT_TREE_ADDR_TREE_0_BASE              0x01000
-#define XPREDICT_PREDICTION_ADDR_PREDICTION_DATA    0x10
-#define XPREDICT_FEATURES_ADDR_FEATURES_0_DATA      0x010
-#define XPREDICT_CONTROL_ADDR_AP_CTRL               0x0
-
-#define DMA_STATUS  0x4      /*R*/
-#define DMA_SA      0x18     /*R-W*/   
-#define DMA_DA      0x20     /*R-W*/
-#define DMA_BTT     0x28     /*R-W*/
-
-#define DMA_ADDRES          0x00000 + XPREDICT_CONTROL_ADDR_AP_CTRL
-#define CONTROL_ADDRES      0x10000 + XPREDICT_CONTROL_ADDR_AP_CTRL
-#define FEATURES_ADDR       0x20000 + XPREDICT_FEATURES_ADDR_FEATURES_0_DATA
-#define PREDICTION_ADDRES   0x30000 + XPREDICT_PREDICTION_ADDR_PREDICTION_DATA
-#define TREES_ADDR          0X80000000 + XPREDICT_TREE_ADDR_TREE_0_BASE
-#define TREES_ADDR_DDR3     0X00000000
