--- linux_drivers/burst_mode/defines_burst.h
@@ -1,18 +1,0 @@
-#define MAP_SIZE 0x20000
-
-#define PING 0xffffffff
-#define PONG 0x00000000
-
-#define FEATURES_PING_ADDR 0xC0000000
-#define FEATURES_PONG_ADDR 0xC2000000
-#define PREDICTIONS_PING_ADDR 0xC4000000
-#define PREDICTIONS_PONG_ADDR 0xC6000000
-
-#define TREES_ADDR 0x00800
-
-#define CONTROL_ADDR_BASE 0x0
-#define CONTROL_ADDR CONTROL_ADDR_BASE + 0
-#define BURST_LENGTH_ADDR CONTROL_ADDR_BASE + 0x10
-#define PING_PONG_ADDR CONTROL_ADDR_BASE + 0x18
-
-#define MAX_BURST 256
