--- linux_drivers/dma/defines_dma.h
@@ -1,17 +1,0 @@
-#define MAP_SIZE 0x20000
-
-#define BASE_ADDRES 0
-
-#define XPREDICT_TREE_ADDR_TREE_0_BASE  0x01000
-#define XPREDICT_PREDICTION_ADDR_PREDICTION_DATA 0x10
-#define XPREDICT_FEATURES_ADDR_FEATURES_0_DATA  0x010
-#define XPREDICT_CONTROL_ADDR_AP_CTRL 0x0
-
-#define CONTROL_ADDRES    0x00000 + XPREDICT_CONTROL_ADDR_AP_CTRL
-#define PREDICTION_ADDRES 0x10000 + XPREDICT_PREDICTION_ADDR_PREDICTION_DATA
-
-#define TREES_ADDR    0x000000 + XPREDICT_TREE_ADDR_TREE_0_BASE
-#define FEATURES_ADDR 0x200000 + XPREDICT_FEATURES_ADDR_FEATURES_0_DATA
-
-#define TREES_SIZE    0x2000  
-#define FEATURES_SIZE 32*8
