--- linux_drivers/memory_mapped/defines_memoty_mapped.h
@@ -1,12 +1,0 @@
-#define BASE_ADDRES 0
-#define MAP_SIZE 0x4000000
-
-#define XPREDICT_TREE_ADDR_TREE_0_BASE  0x01000
-#define XPREDICT_PREDICTION_ADDR_PREDICTION_DATA 0x10
-#define XPREDICT_FEATURES_ADDR_FEATURES_0_DATA  0x010
-#define XPREDICT_CONTROL_ADDR_AP_CTRL 0x0
-
-#define TREE_ADDR         0x000000 + XPREDICT_TREE_ADDR_TREE_0_BASE
-#define CONTROL_ADDRES    0x200000 + XPREDICT_CONTROL_ADDR_AP_CTRL
-#define PREDICTION_ADDRES 0x400000 + XPREDICT_PREDICTION_ADDR_PREDICTION_DATA
-#define FEATURES_ADDRES   0x600000 + XPREDICT_FEATURES_ADDR_FEATURES_0_DATA
